Experiment tracking/mlflow-client-model/requirements.txt
mlflow==2.4.1
scikit-learn==1.2.2
xgboost==1.7.6
matplotlib==3.7.1
seaborn==0.12.2
pandas==2.0.2
numpy==1.24.3
sklearn-pandas==2.2.0
sklearn_pandas==2.2.0
dvc==2.58.2